Add 64-bit integer literal support with intelligent type inference
This commit enables the lexer and parser to handle 64-bit integer literals
directly, removing the previous limitation where only 32-bit values could be
used as literals.

Changes:
- Updated Token and AST structures to use int64_t for integer values
- Modified lexer to use strtoll() for parsing 64-bit integers
- Enhanced runtime to intelligently infer i32 vs i64 based on value range:
  * Values in i32 range (-2147483648 to 2147483647): infer as i32
  * Values outside i32 range but within i64 range: infer as i64
- Fixed unary negation to preserve original integer types (i8-i64, u8-u64)
- Fixed binary arithmetic operations to avoid truncation:
  * Operations now extract values according to promoted type
  * Each integer type (i8-i64, u8-u64) is handled separately
  * Eliminates overflow from casting to i32 prematurely
- Updated documentation to reflect new literal inference behavior

Fixes:
- Large literals like 5000000000 now parse correctly as i64
- Negation of large i64 values preserves type and doesn't overflow
- Arithmetic on i64 values (e.g., 1000000000 + 2000000000) returns correct
  i64 results instead of overflowing to i32

Testing:
- All 221 tests pass
- Verified with test cases for large literals and i64 arithmetic

+++ include/ast.h
@@ -1,6 +1,8 @@
 #ifndef HEMLOCK_AST_H
 #define HEMLOCK_AST_H
 
+#include <stdint.h>  // For int64_t
+
 // Forward declarations
 typedef struct Expr Expr;
 typedef struct Stmt Stmt;
@@ -58,7 +60,7 @@ struct Expr {
     ExprType type;
     union {
         struct {           // ← Number can be int or float
-            int int_value;
+            int64_t int_value;  // Changed to int64_t to support 64-bit literals
             double float_value;
             int is_float;  // flag: which one to use
         } number;
@@ -294,7 +296,7 @@ struct Stmt {
 
 // Expression constructors
 Expr* expr_number(int value);
-Expr* expr_number_int(int value);
+Expr* expr_number_int(int64_t value);  // Changed to int64_t to support 64-bit literals
 Expr* expr_number_float(double value);
 Expr* expr_bool(int value);
 Expr* expr_string(const char *str);
